hosts: add binary-cache config
This exposes a read-only HTTP webserver for the contents in the storage
container.
`rclone serve http` takes care of exposing the storage container over
HTTP.

We disallow listing (by only allowing access to certain paths), and
expose it over HTTP(S) with auto-ssl via caddy.

This will work with whatever domain we route to it, so it's not part of
the configuration.

Signed-off-by: Florian Klink <flokli@flokli.de>