azure-common: support timeout in disk_setup
We need to use cloud-init to format and mount data volumes in azure, we
can't use systemd for it.

Due to
https://github.com/hashicorp/terraform-provider-azurerm/issues/6117,
disks in Azure gets attached late at boot, so any dev-disk-by-….device
units created via systemd-fstab-generator might not exist yet at the
time the graph for multi-user.target is created, causing systemd to fail
starting downstream services due to a missing dependency.

Once the volume is attached, the .device unit pops up via udev, and then
a manual restart of services depending on data disks would work, but
it's messy.

Letting cloud-init take care of data disk mounting (and formatting) is
the right choice, that way systemd doesn't need to do any dependency
tracking of it.

Signed-off-by: Florian Klink <flokli@flokli.de>