azurerm-linux-vm: use azurerm_virtual_machine
We have to fall back to using this resource until
https://github.com/hashicorp/terraform-provider-azurerm/issues/6117
is fixed.

With `azurerm_linux_virtual_machine` and
`azurerm_virtual_machine_data_disk_attachment` the disk only gets
attached once the VM is booted up, and the VM can't boot up if it waits
for the data disk to appear.

Signed-off-by: Florian Klink <flokli@flokli.de>