terraform/jenkins: add post-build-hook and signing
This creates a Nix signing key, and uses terraform-provider-secret to
hold it in the terraform state.

It's then uploaded into an Azure key vault.

The jenkins-controller VM has access to it, and puts it at /etc/secrets/
nix-signing-key.

A post-build-hook is configured, uploading every build to the binary
cache bucket, with the signature.

Signed-off-by: Florian Klink <flokli@flokli.de>