terraform: use custom-nixos.nix to specialize conf
Rather than terraform calling default.nix in the repo root, we call a
little `custom-nixos.nix` trampoline, which calls the flake, then uses
flake.lib.mkNixOS and the systemName to retrieve the azure image.

This script allows two additional arguments passed in as strings,
`extraNixPublicKey` and `extraNixSubstituter`, which we can set through
`nix_argstr` when calling the `azurerm-nix-vm-image` terraform module.

Signed-off-by: Florian Klink <flokli@flokli.de>