configure binary_cache_url
Calculate the https:// binary cache URL for all environments in main.tf,
and use it in the only place it's (currently) used - the config for
caddy (where we previously did the calculation).

We might want to also access that same variable for rendering the Ubuntu
configuration.

The NixOS machines (currently) all run rclone-http, so they can be
configured to substitute from there directly, in the same file that also
enables rclone-http. Refactoring the rclone-http into a module (it needs
to be configurable w.r.t whether read/write is allowed) is left for a
later PR.

Signed-off-by: Florian Klink <flokli@flokli.de>