az-jenkins-controller: add jenkins-upload-artifacts
Provides a Webserver forwarding Multipart uploads to WebDAV.

It acts as glue code between the
[Jenkins HTTP POST plugin](https://plugins.jenkins.io/http-post/) and
`rclone serve webdav`, which can expose many different storage backends.

The WebDAV endpoint to which to forward to is configurable, as well as the
rules to apply in determining the desired target path.

The Jenkins Plugin sends three HTTP Headers (Job Name, Build Number
and Timestamp).
These can be used in a template string, passed as `-target-tpl` argument
to construct the target path.
The service relies on being started through systemd socket activation,
or picks a random port on http to listen on.

Signed-off-by: Florian Klink <flokli@flokli.de>