terraform: Allow changing tfstate deploy location
Deploy the terraform state file on the location declared during
`terraform-init.sh`. Before this change, the terraform state file was
always deployed to North Europe, even if the rest of the infra would
have been deployed to different location.

Also, change `terraform-init.sh` so that it automatically switches to
a private workspace - creating it if needed - to make the script
usage less confusing.

Signed-off-by: Henri Rosten <henri.rosten@unikie.com>