hosts/testagent: downgrade brainstem to 2.9.25
It seems 2.10.5 is unable to change the upstream port:

```
AcronameHubCLI -u 1
Sucessfully connected to USBHub3p SN: 0xEE92E4FD
There was an error (2) changing the upstream port
```

Revert to 2.9.25, which doesn't expose this issue.

Unfortunately, upstream doesn't seem to host that version on their
website anymore (?), or at least the URL looks different, so use the one
we vendored into the `ci-test-automation` repo for now.

Signed-off-by: Florian Klink <flokli@flokli.de>