jenkins-controller: shorten rclone dir-cache-time
We don't want to cache directory listings for 5 minutes (the default),
as that will mean artifacts won't be visible for way too long.

Set to 5 secs, which will still give us some request deduplication, but
not too much.

We cannot apply this to the binary cache configs yet, as rclone seems to
always want to fetch a listing, which is very slow for that.

Signed-off-by: Florian Klink <flokli@flokli.de>