Remove azure builders from dev and priv
Azure builders are no longer needed on 'priv' and 'dev' environments
after https://github.com/tiiuae/ghaf-infra/pull/173.

Also moves the external builder configuration to the
environment-specific configuration options to anticipate ghaf-infra
deployments on locations that might not want to use the
vedenemo.dev builders we are using with Ghaf.

Signed-off-by: Henri Rosten <henri.rosten@unikie.com>