terraform-azurerm: Use azurerm v3.9.85
Use azurerm version v3.9.85 which is the current latest version not
impacted by the issue described in:
https://github.com/hashicorp/terraform-provider-azurerm/issues/24444.

As soon as there's an azurerm upstream fix for the issue, we should
move the override to that version. The override will no longer be needed
when the azurerm version we get from the nixpkgs pinned in ghaf-infra
ships with azurerm version that includes the fix.

Signed-off-by: Henri Rosten <henri.rosten@unikie.com>