binary-cache: use object_store BufReader
Apparently Azure kicks off clients connected for longer than 30 seconds,
and if we have a client needing more than 30 seconds to download a
(large) NAR file, that's a problem.

Generally, we also might want to be a bit more resilient to network
connectivity problems.

Use object_store::buffered::BufReader, rather than directly connecting
the bytes stream from the object storage to the HTTP body.

This keeps the connection to Azure short, by only filling a buffer for
it, and issues new requests once the buffer is drained.

This BufReader also implements AsyncSeek, so we can use the `axum_range`
crate to expose seeking to the client too.
This allows clients to resume downloads if they had connection drops
too.

Signed-off-by: Florian Klink <flokli@flokli.de>