terraform/prod-eun: init terragrunt
This adds an experiment using Terragrunt for managing Terraform
resources and dependencies.

For now, this is limited to the "prod-eun" environment. reusing an
existing state bucket.

It creates a resource group in one state, and then creates a DNS Zone in
another state (prod-eun.az.vedenemo.dev), passing down the resource
group name as an input.

A followup commit will add another state, holding the az.vedenemo.dev
DNS Zone and manage the delegation of prod-eun.az.vedenemo.dev.

Signed-off-by: Florian Klink <flokli@flokli.de>