hosts/azure: add users to wheel, make trusted
Use membership in the "wheel" group to control whether users are allowed
to sudo or not, rather than managing it explicitly.

Also configure everyone in the wheel group to be a trusted Nix user.
These same users where already able to sudo without password, so this
doesn't give them more privileges.

However, it allows them to also use nix-copy-closure to upload
loacally-built closures to a system, which is a great way to
incrementally test changes without having to recreate a new VM image
every time.

A subsequent terraform run would still redeploy a new VM with a new
image, but this can dramatically reduce the iteration time during
development.

Signed-off-by: Florian Klink <flokli@flokli.de>