az-jenkins-controller: drop jenkins-config unit
The only thing this systemd service still did was disabling the initial
setup, by running a groovy script (via jenkins-cli), and then triggering
a Jenkins reboot.

As we don't need the initial password anymore, we can simply disable the
setup wizard entirely, and remove the need for that unit alltogether.

Signed-off-by: Florian Klink <flokli@flokli.de>