Check disk availability before flash (#677)
After mounting the target disk to test agent
it has not been checked if the disk is actually
visible on test agent. Sometimes random failure
(in USB hub?) in getting USB SSD connected has
led to repeating errors because if the target
disk is not available flash stage will dd the
image to a normal file (with the same name as the
target disk). Once that file is created it will
prevent all further flashing attempt to the disk
with the same name.

Now check if the target disk is available and if
only a normal file with the same name is found
delete that file.

Signed-off-by: Samuli Leivo <samuli.leivo@unikie.com>